+++ saritasa_drf_tools/serializers/mixins/__init__.py
@@ -1,3 +1,4 @@
 from .clean_validation_mixin import CleanValidationMixin
 from .field_mapping_override_mixin import FieldMappingOverride
+from .nested_fields_mixin import NestedFieldsMixin
 from .user_and_request_from_request_mixin import UserAndRequestFromContextMixin